Move calculate_reachability to attackgraph.query
