+++ maltoolbox/attackgraph/attackgraph.py
@@ -686,75 +686,3 @@ class AttackGraph():
         if not isinstance(attacker.id, int):
             raise ValueError(f'Invalid attacker id.')
         del self._id_to_attacker[attacker.id]
-
-    def calculate_reachability_for_attacker(self, attacker: Attacker) -> None:
-        """Calculate the reachability for an attacker to all attack steps
-        starting from the reached attack steps and propagating to descendants
-
-        Args:
-        attacker - the attacker to calculate reachability for
-        """
-
-        def attack_step_reachable_by(
-                node_with_reachable_parent: AttackGraphNode,
-                attacker: Attacker
-            ) -> bool:
-            """
-            Decide if attack step node with at least one reachable parent
-            is reachable by attacker based on its viability, type and parents
-            """
-
-            if not node_with_reachable_parent.is_viable:
-                return False
-
-            if node_with_reachable_parent.type == "or":
-                # OR-Node is reachable if any parent is reachable
-                return True
-            if node_with_reachable_parent.type == "and":
-                # Node is reachable only if all parents are reachable
-                return all(
-                    parent.is_reachable_by(attacker)
-                    for parent in node_with_reachable_parent.parents
-                )
-
-            # Not an attack step -> Not reachable
-            return False
-
-        def propagate_reachable(
-            reachable_node: AttackGraphNode, attacker: Attacker) -> None:
-            """
-            Mark node reachable for attacker and propagate to 
-            reachable children recursively.
-            """
-
-            if attacker in reachable_node.reachable_by:
-                # Already visited/calculated for attacker
-                return
-
-            reachable_node.reachable_by.add(attacker)
-            attacker.reachable_attack_steps.add(reachable_node)
-
-            for child in reachable_node.children:
-                if attack_step_reachable_by(child, attacker):
-                    propagate_reachable(child, attacker)
-
-        # All reached attack_steps are reachable,
-        # propagate reachability to their children
-        for reached_node in attacker.reached_attack_steps:
-            propagate_reachable(reached_node, attacker)
-
-    def calculate_reachability(self) -> None:
-        """Mark reachable nodes reachable by each attacker
-
-        For each attacker, calculate which nodes they can reach and
-        update values in node.reachable_by and attacker.reachable_attack_steps
-        """
-
-        for node in self.nodes:
-            # clear reachable_by for each node
-            node.reachable_by.clear()
-
-        for attacker in self.attackers:
-            # clear attacker reachable attack steps
-            attacker.reachable_attack_steps.clear()
-            self.calculate_reachability_for_attacker(attacker)
